+++ src/test/java/com/astrapay/qris/QrisMapperTest.java
@@ -1,16 +1,11 @@
 package com.astrapay.qris;
 
+import com.astrapay.qris.mpm.QrisParser;
 import com.astrapay.qris.object.*;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import java.lang.reflect.Field;
-import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.mock;
